refactor(remote): class-based remote helpers
